+++ Lyubarsky/KP8/include/Node.hpp
@@ -1,0 +1,13 @@
+#ifndef NODE_HPP
+#define NODE_HPP
+
+template<typename T>
+struct node{
+    node<T>* nextNode;
+    T data;
+    node(const T& = T(), node<T>* = nullptr);
+};
+
+#include "../src/Node.cpp"
+
+#endif
